requirements.txt
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.3
cycler==0.10.0
decorator==4.4.2
grpcio==1.37.1
grpcio-tools==1.37.1
idna==3.2
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.21.1
Pillow==8.3.1
protobuf==3.17.3
pycodestyle==2.7.0
pymilvus==1.1.2
pyparsing==2.4.7
python-dateutil==2.8.2
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
threadpoolctl==2.2.0
toml==0.10.2
tqdm==4.61.2
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.6
